# docs/archive/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "World Weaver Memory Gateway API",
    "version": "1.0.0",
    "description": "MCP server providing access to a tripartite memory system (episodic, semantic, procedural) for Claude Code instances. Implements 17 tools for autobiographical events, knowledge graphs, and learned procedures with ACT-R activation, FSRS retrievability, and Hebbian learning.",
    "contact": {
      "name": "World Weaver Project",
      "url": "https://github.com/astoreyai/world-weaver"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local MCP server"
    }
  ],
  "paths": {
...
    "/tools/create_entity": {
      "post": {
        "summary": "Create a semantic knowledge entity",
        "description": "Creates a new entity in the knowledge graph",
        "operationId": "create_entity",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Canonical entity name"
                  },
                  "entity_type": {
                    "type": "string",
                    "enum": [
                      "CONCEPT",
                      "PERSON",
                      "PROJECT",
                      "TOOL",
                      "TECHNIQUE",
                      "FACT"
                    ]
                  },
                  "summary": {
                    "type": "string",
                    "description": "Short description"
                  },
                  "details": {
                    "type": "string",
                    "description": "Expanded context"
                  },
                  "source": {
                    "type": "string",
                    "description": "Episode ID or 'user_provided'"
                  }
                },
                "required": [
                  "name",
                  "entity_type",
                  "summary"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "field": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/tools/create_relation": {
      "post": {
        "summary": "Create a Hebbian-weighted relationship between entities",
        "description": "Links two entities with a relationship type and initial weight",
        "operationId": "create_relation",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "source_id": {
                    "type": "string",
                    "description": "Source entity UUID"
                  },
                  "target_id": {
                    "type": "string",
                    "description": "Target entity UUID"
                  },
                  "relation_type": {
                    "type": "string",
                    "enum": [
                      "USES",
                      "PRODUCES",
                      "REQUIRES",
                      "CAUSES",
                      "PART_OF",
                      "SIMILAR_TO",
                      "IMPLEMENTS"
                    ]
                  },
                  "initial_weight": {
                    "type": "number",
                    "default": 0.1,
                    "minimum": 0.0,
                    "maximum": 1.0
                  }
                },
                "required": [
                  "source_id",
                  "target_id",
                  "relation_type"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "field": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/tools/semantic_recall": {
      "post": {
        "summary": "Retrieve semantic entities with ACT-R activation scoring",
        "description": "Combines vector similarity + ACT-R activation + FSRS retrievability",
        "operationId": "semantic_recall",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "query": {
                    "type": "string",
                    "description": "Natural language search query"
                  },
                  "context_entities": {
                    "type": "array",
                    "description": "Entity IDs to spread activation from",
                    "items": {
                      "type": "string"
                    }
                  },
                  "limit": {
                    "type": "integer",
                    "default": 10
                  },
                  "offset": {
                    "type": "integer",
                    "description": "Number of results to skip (pagination)",
                    "default": 0
                  },
                  "include_spreading": {
                    "type": "boolean",
                    "default": true
                  }
                },
                "required": [
                  "query"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "field": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/tools/spread_activation": {
      "post": {
        "summary": "Spread activation through knowledge graph from seed entities",
        "description": "Propagates activation through the graph to find related concepts",
        "operationId": "spread_activation",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "seed_entities": {
                    "type": "array",
                    "description": "Entity IDs to start from",
                    "items": {
                      "type": "string"
                    }
                  },
                  "steps": {
                    "type": "integer",
                    "description": "Number of propagation steps",
                    "default": 3
                  },
                  "retention": {
                    "type": "number",
                    "default": 0.5,
                    "minimum": 0.0,
                    "maximum": 1.0
                  },
                  "decay": {
                    "type": "number",
                    "default": 0.1,
                    "minimum": 0.0,
                    "maximum": 1.0
                  },
                  "threshold": {
                    "type": "number",
                    "default": 0.01,
                    "minimum": 0.0,
                    "maximum": 1.0
                  }
                },
                "required": [
                  "seed_entities"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "field": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}